+++ tests/unit/test_compiler.py
@@ -21,7 +21,28 @@ import pytest
 import sqlalchemy.exc
 
 from .conftest import setup_table
-from .conftest import sqlalchemy_1_4_or_higher, sqlalchemy_before_1_4
+from .conftest import (
+    sqlalchemy_2_0_or_higher,
+    sqlalchemy_before_2_0,
+)
+from sqlalchemy.sql.functions import rollup, cube, grouping_sets
+
+
+@pytest.fixture
+def table(faux_conn, metadata):
+    # Fixture to create a sample table for testing
+
+    table = setup_table(
+        faux_conn,
+        "table1",
+        metadata,
+        sqlalchemy.Column("foo", sqlalchemy.Integer),
+        sqlalchemy.Column("bar", sqlalchemy.ARRAY(sqlalchemy.Integer)),
+    )
+
+    yield table
+
+    table.drop(faux_conn)
 
 
 def test_constraints_are_ignored(faux_conn, metadata):
@@ -58,7 +79,6 @@ def test_cant_compile_unnamed_column(faux_conn, metadata):
         sqlalchemy.Column(sqlalchemy.Integer).compile(faux_conn)
 
 
-@sqlalchemy_1_4_or_higher
 def test_no_alias_for_known_tables(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/353
     table = setup_table(
@@ -80,7 +100,6 @@ def test_no_alias_for_known_tables(faux_conn, metadata):
     assert found_sql == expected_sql
 
 
-@sqlalchemy_1_4_or_higher
 def test_no_alias_for_known_tables_cte(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
     table = setup_table(
@@ -142,30 +161,37 @@ def prepare_implicit_join_base_query(
     return q
 
 
-@sqlalchemy_before_1_4
-def test_no_implicit_join_asterix_for_inner_unnest_before_1_4(faux_conn, metadata):
-    # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
-    q = prepare_implicit_join_base_query(faux_conn, metadata, True, True)
-    expected_initial_sql = (
-        "SELECT `table1`.`foo`, `table2`.`bar` \n"
-        "FROM `table2`, unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`"
+# Test vendored method update_from_clause()
+# from sqlalchemy_bigquery_vendored.sqlalchemy.postgresql.base.PGCompiler
+def test_update_from_clause(faux_conn, metadata):
+    table1 = setup_table(
+        faux_conn,
+        "table1",
+        metadata,
+        sqlalchemy.Column("foo", sqlalchemy.String),
+        sqlalchemy.Column("bar", sqlalchemy.Integer),
     )
-    found_initial_sql = q.compile(faux_conn).string
-    assert found_initial_sql == expected_initial_sql
-
-    q = sqlalchemy.select(["*"]).select_from(q)
-
-    expected_outer_sql = (
-        "SELECT * \n"
-        "FROM (SELECT `table1`.`foo` AS `foo`, `table2`.`bar` AS `bar` \n"
-        "FROM `table2`, unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`)"
+    table2 = setup_table(
+        faux_conn,
+        "table2",
+        metadata,
+        sqlalchemy.Column("foo", sqlalchemy.String),
+        sqlalchemy.Column("bar", sqlalchemy.Integer),
     )
-    found_outer_sql = q.compile(faux_conn).string
-    assert found_outer_sql == expected_outer_sql
+
+    stmt = (
+        sqlalchemy.update(table1)
+        .where(table1.c.foo == table2.c.foo)
+        .where(table2.c.bar == 1)
+        .values(bar=2)
+    )
+    expected_sql = "UPDATE `table1` SET `bar`=%(bar:INT64)s FROM `table2` WHERE `table1`.`foo` = `table2`.`foo` AND `table2`.`bar` = %(bar_1:INT64)s"
+    found_sql = stmt.compile(faux_conn).string
+    assert found_sql == expected_sql
 
 
-@sqlalchemy_1_4_or_higher
-def test_no_implicit_join_asterix_for_inner_unnest(faux_conn, metadata):
+@sqlalchemy_before_2_0
+def test_no_implicit_join_asterix_for_inner_unnest_before_2_0(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
     q = prepare_implicit_join_base_query(faux_conn, metadata, True, False)
     expected_initial_sql = (
@@ -187,30 +213,31 @@ def test_no_implicit_join_asterix_for_inner_unnest(faux_conn, metadata):
     assert found_outer_sql == expected_outer_sql
 
 
-@sqlalchemy_before_1_4
-def test_no_implicit_join_for_inner_unnest_before_1_4(faux_conn, metadata):
+@sqlalchemy_2_0_or_higher
+def test_no_implicit_join_asterix_for_inner_unnest(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
-    q = prepare_implicit_join_base_query(faux_conn, metadata, True, True)
+    q = prepare_implicit_join_base_query(faux_conn, metadata, True, False)
     expected_initial_sql = (
         "SELECT `table1`.`foo`, `table2`.`bar` \n"
-        "FROM `table2`, unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`"
+        "FROM unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`, `table2`"
     )
     found_initial_sql = q.compile(faux_conn).string
     assert found_initial_sql == expected_initial_sql
 
-    q = sqlalchemy.select([q.c.foo]).select_from(q)
+    q = q.subquery()
+    q = sqlalchemy.select("*").select_from(q)
 
     expected_outer_sql = (
-        "SELECT `foo` \n"
+        "SELECT * \n"
         "FROM (SELECT `table1`.`foo` AS `foo`, `table2`.`bar` AS `bar` \n"
-        "FROM `table2`, unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`)"
+        "FROM unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`, `table2`) AS `anon_1`"
     )
     found_outer_sql = q.compile(faux_conn).string
     assert found_outer_sql == expected_outer_sql
 
 
-@sqlalchemy_1_4_or_higher
-def test_no_implicit_join_for_inner_unnest(faux_conn, metadata):
+@sqlalchemy_before_2_0
+def test_no_implicit_join_for_inner_unnest_before_2_0(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
     q = prepare_implicit_join_base_query(faux_conn, metadata, True, False)
     expected_initial_sql = (
@@ -232,7 +259,29 @@ def test_no_implicit_join_for_inner_unnest(faux_conn, metadata):
     assert found_outer_sql == expected_outer_sql
 
 
-@sqlalchemy_1_4_or_higher
+@sqlalchemy_2_0_or_higher
+def test_no_implicit_join_for_inner_unnest(faux_conn, metadata):
+    # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
+    q = prepare_implicit_join_base_query(faux_conn, metadata, True, False)
+    expected_initial_sql = (
+        "SELECT `table1`.`foo`, `table2`.`bar` \n"
+        "FROM unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`, `table2`"
+    )
+    found_initial_sql = q.compile(faux_conn).string
+    assert found_initial_sql == expected_initial_sql
+
+    q = q.subquery()
+    q = sqlalchemy.select(q.c.foo).select_from(q)
+
+    expected_outer_sql = (
+        "SELECT `anon_1`.`foo` \n"
+        "FROM (SELECT `table1`.`foo` AS `foo`, `table2`.`bar` AS `bar` \n"
+        "FROM unnest(`table2`.`foos`) AS `unnested_foos` JOIN `table1` ON `table1`.`foo` = `unnested_foos`, `table2`) AS `anon_1`"
+    )
+    found_outer_sql = q.compile(faux_conn).string
+    assert found_outer_sql == expected_outer_sql
+
+
 def test_no_implicit_join_asterix_for_inner_unnest_no_table2_column(
     faux_conn, metadata
 ):
@@ -257,7 +306,6 @@ def test_no_implicit_join_asterix_for_inner_unnest_no_table2_column(
     assert found_outer_sql == expected_outer_sql
 
 
-@sqlalchemy_1_4_or_higher
 def test_no_implicit_join_for_inner_unnest_no_table2_column(faux_conn, metadata):
     # See: https://github.com/googleapis/python-bigquery-sqlalchemy/issues/368
     q = prepare_implicit_join_base_query(faux_conn, metadata, False, False)
@@ -278,3 +326,133 @@ def test_no_implicit_join_for_inner_unnest_no_table2_column(faux_conn, metadata)
     )
     found_outer_sql = q.compile(faux_conn).string
     assert found_outer_sql == expected_outer_sql
+
+
+grouping_ops = (
+    "grouping_op, grouping_op_func",
+    [("GROUPING SETS", grouping_sets), ("ROLLUP", rollup), ("CUBE", cube)],
+)
+
+
+@pytest.mark.parametrize(*grouping_ops)
+def test_grouping_ops_vs_single_column(faux_conn, table, grouping_op, grouping_op_func):
+    # Tests each of the grouping ops against a single column
+
+    q = sqlalchemy.select(table.c.foo).group_by(grouping_op_func(table.c.foo))
+    found_sql = q.compile(faux_conn).string
+
+    expected_sql = (
+        f"SELECT `table1`.`foo` \n"
+        f"FROM `table1` GROUP BY {grouping_op}(`table1`.`foo`)"
+    )
+
+    assert found_sql == expected_sql
+
+
+@pytest.mark.parametrize(*grouping_ops)
+def test_grouping_ops_vs_multi_columns(faux_conn, table, grouping_op, grouping_op_func):
+    # Tests each of the grouping ops against multiple columns
+
+    q = sqlalchemy.select(table.c.foo, table.c.bar).group_by(
+        grouping_op_func(table.c.foo, table.c.bar)
+    )
+    found_sql = q.compile(faux_conn).string
+
+    expected_sql = (
+        f"SELECT `table1`.`foo`, `table1`.`bar` \n"
+        f"FROM `table1` GROUP BY {grouping_op}(`table1`.`foo`, `table1`.`bar`)"
+    )
+
+    assert found_sql == expected_sql
+
+
+@pytest.mark.parametrize(*grouping_ops)
+def test_grouping_op_with_grouping_op(faux_conn, table, grouping_op, grouping_op_func):
+    # Tests multiple grouping ops in a single statement
+
+    q = sqlalchemy.select(table.c.foo, table.c.bar).group_by(
+        grouping_op_func(table.c.foo, table.c.bar), grouping_op_func(table.c.foo)
+    )
+    found_sql = q.compile(faux_conn).string
+
+    expected_sql = (
+        f"SELECT `table1`.`foo`, `table1`.`bar` \n"
+        f"FROM `table1` GROUP BY {grouping_op}(`table1`.`foo`, `table1`.`bar`), {grouping_op}(`table1`.`foo`)"
+    )
+
+    assert found_sql == expected_sql
+
+
+@pytest.mark.parametrize(*grouping_ops)
+def test_grouping_ops_vs_group_by(faux_conn, table, grouping_op, grouping_op_func):
+    # Tests grouping op against regular group by statement
+
+    q = sqlalchemy.select(table.c.foo, table.c.bar).group_by(
+        table.c.foo, grouping_op_func(table.c.bar)
+    )
+    found_sql = q.compile(faux_conn).string
+
+    expected_sql = (
+        f"SELECT `table1`.`foo`, `table1`.`bar` \n"
+        f"FROM `table1` GROUP BY `table1`.`foo`, {grouping_op}(`table1`.`bar`)"
+    )
+
+    assert found_sql == expected_sql
+
+
+@pytest.mark.parametrize(*grouping_ops)
+def test_complex_grouping_ops_vs_nested_grouping_ops(
+    faux_conn, table, grouping_op, grouping_op_func
+):
+    # Tests grouping ops nested within grouping ops
+
+    q = sqlalchemy.select(table.c.foo, table.c.bar).group_by(
+        grouping_sets(table.c.foo, grouping_op_func(table.c.bar))
+    )
+    found_sql = q.compile(faux_conn).string
+
+    expected_sql = (
+        f"SELECT `table1`.`foo`, `table1`.`bar` \n"
+        f"FROM `table1` GROUP BY GROUPING SETS(`table1`.`foo`, {grouping_op}(`table1`.`bar`))"
+    )
+
+    assert found_sql == expected_sql
+
+
+def test_label_compiler(faux_conn, metadata):
+    class CustomLower(sqlalchemy.sql.functions.FunctionElement):
+        name = "custom_lower"
+
+    @sqlalchemy.ext.compiler.compiles(CustomLower)
+    def compile_custom_intersect(element, compiler, **kwargs):
+        if compiler.dialect.name != "bigquery":
+            # We only test with the BigQuery dialect, so this should never happen.
+            raise sqlalchemy.exc.CompileError(  # pragma: NO COVER
+                f"custom_lower is not supported for dialect {compiler.dialect.name}"
+            )
+
+        clauses = list(element.clauses)
+        field = compiler.process(clauses[0], **kwargs)
+        return f"LOWER({field})"
+
+    table1 = setup_table(
+        faux_conn,
+        "table1",
+        metadata,
+        sqlalchemy.Column("foo", sqlalchemy.String),
+        sqlalchemy.Column("bar", sqlalchemy.Integer),
+    )
+
+    lower_foo = CustomLower(table1.c.foo).label("some_label")
+    q = (
+        sqlalchemy.select(lower_foo, sqlalchemy.func.max(table1.c.bar))
+        .select_from(table1)
+        .group_by(lower_foo)
+    )
+    expected_sql = (
+        "SELECT LOWER(`table1`.`foo`) AS `some_label`, max(`table1`.`bar`) AS `max_1` \n"
+        "FROM `table1` GROUP BY `some_label`"
+    )
+
+    found_sql = q.compile(faux_conn).string
+    assert found_sql == expected_sql