+++ tests/unit/test_select.py
@@ -20,25 +20,18 @@
 import datetime
 from decimal import Decimal
 
-import packaging.version
 import pytest
 import sqlalchemy
 from sqlalchemy import not_
 
 import sqlalchemy_bigquery
 
-from .conftest import (
-    setup_table,
-    sqlalchemy_version,
-    sqlalchemy_1_3_or_higher,
-    sqlalchemy_1_4_or_higher,
-    sqlalchemy_before_1_4,
-)
+from .conftest import setup_table
 
 
 def test_labels_not_forced(faux_conn):
     table = setup_table(faux_conn, "t", sqlalchemy.Column("id", sqlalchemy.Integer))
-    result = faux_conn.execute(sqlalchemy.select([table.c.id]))
+    result = faux_conn.execute(sqlalchemy.select(table.c.id))
     assert result.keys() == ["id"]  # Look! Just the column name!
 
 
@@ -154,14 +147,18 @@ def test_typed_parameters(faux_conn, type_, val, btype, vrep):
         {},
     )
 
-    assert list(map(list, faux_conn.execute(sqlalchemy.select([table])))) == [[val]] * 2
+    assert list(map(list, faux_conn.execute(sqlalchemy.select(table)))) == [[val]] * 2
     assert faux_conn.test_data["execute"][-1][0] == "SELECT `t`.`foo` \nFROM `t`"
 
     assert (
         list(
             map(
                 list,
-                faux_conn.execute(sqlalchemy.select([table.c.foo], use_labels=True)),
+                faux_conn.execute(
+                    sqlalchemy.select(table.c.foo).set_label_style(
+                        sqlalchemy.LABEL_STYLE_TABLENAME_PLUS_COL
+                    )
+                ),
             )
         )
         == [[val]] * 2
@@ -169,6 +166,94 @@ def test_typed_parameters(faux_conn, type_, val, btype, vrep):
     assert faux_conn.test_data["execute"][-1][0] == (
         "SELECT `t`.`foo` AS `t_foo` \nFROM `t`"
     )
+
+
+def test_except(faux_conn):
+    table = setup_table(
+        faux_conn,
+        "table",
+        sqlalchemy.Column("id", sqlalchemy.Integer),
+        sqlalchemy.Column("foo", sqlalchemy.Integer),
+    )
+
+    s1 = sqlalchemy.select(table.c.foo).where(table.c.id >= 2)
+    s2 = sqlalchemy.select(table.c.foo).where(table.c.id >= 4)
+
+    s3 = s1.except_(s2)
+
+    result = s3.compile(faux_conn).string
+
+    expected = (
+        "SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_1:INT64)s EXCEPT DISTINCT SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_2:INT64)s"
+    )
+    assert result == expected
+
+
+def test_intersect(faux_conn):
+    table = setup_table(
+        faux_conn,
+        "table",
+        sqlalchemy.Column("id", sqlalchemy.Integer),
+        sqlalchemy.Column("foo", sqlalchemy.Integer),
+    )
+
+    s1 = sqlalchemy.select(table.c.foo).where(table.c.id >= 2)
+    s2 = sqlalchemy.select(table.c.foo).where(table.c.id >= 4)
+
+    s3 = s1.intersect(s2)
+
+    result = s3.compile(faux_conn).string
+
+    expected = (
+        "SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_1:INT64)s INTERSECT DISTINCT SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_2:INT64)s"
+    )
+    assert result == expected
+
+
+def test_union(faux_conn):
+    table = setup_table(
+        faux_conn,
+        "table",
+        sqlalchemy.Column("id", sqlalchemy.Integer),
+        sqlalchemy.Column("foo", sqlalchemy.Integer),
+    )
+
+    s1 = sqlalchemy.select(table.c.foo).where(table.c.id >= 2)
+    s2 = sqlalchemy.select(table.c.foo).where(table.c.id >= 4)
+
+    s3 = s1.union(s2)
+
+    result = s3.compile(faux_conn).string
+
+    expected = (
+        "SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_1:INT64)s UNION DISTINCT SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_2:INT64)s"
+    )
+    assert result == expected
+
+    s4 = s1.union_all(s2)
+
+    result = s4.compile(faux_conn).string
+
+    expected = (
+        "SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_1:INT64)s UNION ALL SELECT `table`.`foo` \n"
+        "FROM `table` \n"
+        "WHERE `table`.`id` >= %(id_2:INT64)s"
+    )
+    assert result == expected
 
 
 def test_select_struct(faux_conn, metadata):
@@ -183,7 +268,7 @@ def test_select_struct(faux_conn, metadata):
     faux_conn.ex("create table t (x RECORD)")
     faux_conn.ex("""insert into t values ('{"y": 1}')""")
 
-    row = list(faux_conn.execute(sqlalchemy.select([table])))[0]
+    row = list(faux_conn.execute(sqlalchemy.select(table)))[0]
     # We expect the raw string, because sqlite3, unlike BigQuery
     # doesn't deserialize for us.
     assert row.x == '{"y": 1}'
@@ -191,7 +276,7 @@ def test_select_struct(faux_conn, metadata):
 
 def test_select_label_starts_w_digit(faux_conn):
     # Make sure label names are legal identifiers
-    faux_conn.execute(sqlalchemy.select([sqlalchemy.literal(1).label("2foo")]))
+    faux_conn.execute(sqlalchemy.select(sqlalchemy.literal(1).label("2foo")))
     assert (
         faux_conn.test_data["execute"][-1][0] == "SELECT %(param_1:INT64)s AS `_2foo`"
     )
@@ -205,7 +290,7 @@ def test_force_quote(faux_conn):
         "t",
         sqlalchemy.Column(quoted_name("foo", True), sqlalchemy.Integer),
     )
-    faux_conn.execute(sqlalchemy.select([table]))
+    faux_conn.execute(sqlalchemy.select(table))
     assert faux_conn.test_data["execute"][-1][0] == ("SELECT `t`.`foo` \nFROM `t`")
 
 
@@ -217,26 +302,12 @@ def test_disable_quote(faux_conn):
         "t",
         sqlalchemy.Column(quoted_name("foo", False), sqlalchemy.Integer),
     )
-    faux_conn.execute(sqlalchemy.select([table]))
+    faux_conn.execute(sqlalchemy.select(table))
     assert faux_conn.test_data["execute"][-1][0] == ("SELECT `t`.foo \nFROM `t`")
 
 
-@sqlalchemy_before_1_4
-def test_select_in_lit_13(faux_conn):
-    [[isin]] = faux_conn.execute(
-        sqlalchemy.select([sqlalchemy.literal(1).in_([1, 2, 3])])
-    )
-    assert isin
-    assert faux_conn.test_data["execute"][-1] == (
-        "SELECT %(param_1:INT64)s IN "
-        "(%(param_2:INT64)s, %(param_3:INT64)s, %(param_4:INT64)s) AS `anon_1`",
-        {"param_1": 1, "param_2": 1, "param_3": 2, "param_4": 3},
-    )
-
-
-@sqlalchemy_1_4_or_higher
 def test_select_in_lit(faux_conn, last_query):
-    faux_conn.execute(sqlalchemy.select([sqlalchemy.literal(1).in_([1, 2, 3])]))
+    faux_conn.execute(sqlalchemy.select(sqlalchemy.literal(1).in_([1, 2, 3])))
     last_query(
         "SELECT %(param_1:INT64)s IN UNNEST(%(param_2:INT64)s) AS `anon_1`",
         {"param_1": 1, "param_2": [1, 2, 3]},
@@ -244,83 +315,47 @@ def test_select_in_lit(faux_conn, last_query):
 
 
 def test_select_in_param(faux_conn, last_query):
-    [[isin]] = faux_conn.execute(
+    faux_conn.execute(
         sqlalchemy.select(
-            [sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))]
+            sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))
         ),
         dict(q=[1, 2, 3]),
     )
-    if sqlalchemy_version >= packaging.version.parse("1.4"):
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
-            {"param_1": 1, "q": [1, 2, 3]},
-        )
-    else:
-        assert isin
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST("
-            "[ %(q_1:INT64)s, %(q_2:INT64)s, %(q_3:INT64)s ]"
-            ") AS `anon_1`",
-            {"param_1": 1, "q_1": 1, "q_2": 2, "q_3": 3},
-        )
+
+    last_query(
+        "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
+        {"param_1": 1, "q": [1, 2, 3]},
+    )
 
 
 def test_select_in_param1(faux_conn, last_query):
-    [[isin]] = faux_conn.execute(
+    faux_conn.execute(
         sqlalchemy.select(
-            [sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))]
+            sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))
         ),
         dict(q=[1]),
     )
-    if sqlalchemy_version >= packaging.version.parse("1.4"):
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
-            {"param_1": 1, "q": [1]},
-        )
-    else:
-        assert isin
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST(" "[ %(q_1:INT64)s ]" ") AS `anon_1`",
-            {"param_1": 1, "q_1": 1},
-        )
+    last_query(
+        "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
+        {"param_1": 1, "q": [1]},
+    )
 
 
-@sqlalchemy_1_3_or_higher
 def test_select_in_param_empty(faux_conn, last_query):
-    [[isin]] = faux_conn.execute(
+    faux_conn.execute(
         sqlalchemy.select(
-            [sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))]
+            sqlalchemy.literal(1).in_(sqlalchemy.bindparam("q", expanding=True))
         ),
         dict(q=[]),
     )
-    if sqlalchemy_version >= packaging.version.parse("1.4"):
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
-            {"param_1": 1, "q": []},
-        )
-    else:
-        assert not isin
-        last_query(
-            "SELECT %(param_1:INT64)s IN UNNEST([  ]) AS `anon_1`", {"param_1": 1}
-        )
-
-
-@sqlalchemy_before_1_4
-def test_select_notin_lit13(faux_conn):
-    [[isnotin]] = faux_conn.execute(
-        sqlalchemy.select([sqlalchemy.literal(0).notin_([1, 2, 3])])
-    )
-    assert isnotin
-    assert faux_conn.test_data["execute"][-1] == (
-        "SELECT (%(param_1:INT64)s NOT IN "
-        "(%(param_2:INT64)s, %(param_3:INT64)s, %(param_4:INT64)s)) AS `anon_1`",
-        {"param_1": 0, "param_2": 1, "param_3": 2, "param_4": 3},
+    last_query(
+        "SELECT %(param_1:INT64)s IN UNNEST(%(q:INT64)s) AS `anon_1`",
+        {"param_1": 1, "q": []},
     )
 
 
-@sqlalchemy_1_4_or_higher
 def test_select_notin_lit(faux_conn, last_query):
-    faux_conn.execute(sqlalchemy.select([sqlalchemy.literal(0).notin_([1, 2, 3])]))
+    faux_conn.execute(sqlalchemy.select(sqlalchemy.literal(0).notin_([1, 2, 3])))
     last_query(
         "SELECT (%(param_1:INT64)s NOT IN UNNEST(%(param_2:INT64)s)) AS `anon_1`",
         {"param_1": 0, "param_2": [1, 2, 3]},
@@ -328,45 +363,29 @@ def test_select_notin_lit(faux_conn, last_query):
 
 
 def test_select_notin_param(faux_conn, last_query):
-    [[isnotin]] = faux_conn.execute(
+    faux_conn.execute(
         sqlalchemy.select(
-            [sqlalchemy.literal(1).notin_(sqlalchemy.bindparam("q", expanding=True))]
+            sqlalchemy.literal(1).notin_(sqlalchemy.bindparam("q", expanding=True))
         ),
         dict(q=[1, 2, 3]),
     )
-    if sqlalchemy_version >= packaging.version.parse("1.4"):
-        last_query(
-            "SELECT (%(param_1:INT64)s NOT IN UNNEST(%(q:INT64)s)) AS `anon_1`",
-            {"param_1": 1, "q": [1, 2, 3]},
-        )
-    else:
-        assert not isnotin
-        last_query(
-            "SELECT (%(param_1:INT64)s NOT IN UNNEST("
-            "[ %(q_1:INT64)s, %(q_2:INT64)s, %(q_3:INT64)s ]"
-            ")) AS `anon_1`",
-            {"param_1": 1, "q_1": 1, "q_2": 2, "q_3": 3},
-        )
+    last_query(
+        "SELECT (%(param_1:INT64)s NOT IN UNNEST(%(q:INT64)s)) AS `anon_1`",
+        {"param_1": 1, "q": [1, 2, 3]},
+    )
 
 
-@sqlalchemy_1_3_or_higher
 def test_select_notin_param_empty(faux_conn, last_query):
-    [[isnotin]] = faux_conn.execute(
+    faux_conn.execute(
         sqlalchemy.select(
-            [sqlalchemy.literal(1).notin_(sqlalchemy.bindparam("q", expanding=True))]
+            sqlalchemy.literal(1).notin_(sqlalchemy.bindparam("q", expanding=True))
         ),
         dict(q=[]),
     )
-    if sqlalchemy_version >= packaging.version.parse("1.4"):
-        last_query(
-            "SELECT (%(param_1:INT64)s NOT IN UNNEST(%(q:INT64)s)) AS `anon_1`",
-            {"param_1": 1, "q": []},
-        )
-    else:
-        assert isnotin
-        last_query(
-            "SELECT (%(param_1:INT64)s NOT IN UNNEST([  ])) AS `anon_1`", {"param_1": 1}
-        )
+    last_query(
+        "SELECT (%(param_1:INT64)s NOT IN UNNEST(%(q:INT64)s)) AS `anon_1`",
+        {"param_1": 1, "q": []},
+    )
 
 
 def test_literal_binds_kwarg_with_an_IN_operator_252(faux_conn):
@@ -376,7 +395,7 @@ def test_literal_binds_kwarg_with_an_IN_operator_252(faux_conn):
         sqlalchemy.Column("val", sqlalchemy.Integer),
         initial_data=[dict(val=i) for i in range(3)],
     )
-    q = sqlalchemy.select([table.c.val]).where(table.c.val.in_([2]))
+    q = sqlalchemy.select(table.c.val).where(table.c.val.in_([2]))
 
     def nstr(q):
         return " ".join(str(q).strip().split())
@@ -387,7 +406,6 @@ def test_literal_binds_kwarg_with_an_IN_operator_252(faux_conn):
     )
 
 
-@sqlalchemy_1_4_or_higher
 @pytest.mark.parametrize("alias", [True, False])
 def test_unnest(faux_conn, alias):
     from sqlalchemy import String
@@ -405,7 +423,6 @@ def test_unnest(faux_conn, alias):
     )
 
 
-@sqlalchemy_1_4_or_higher
 @pytest.mark.parametrize("alias", [True, False])
 def test_table_valued_alias_w_multiple_references_to_the_same_table(faux_conn, alias):
     from sqlalchemy import String
@@ -424,7 +441,6 @@ def test_table_valued_alias_w_multiple_references_to_the_same_table(faux_conn, a
     )
 
 
-@sqlalchemy_1_4_or_higher
 @pytest.mark.parametrize("alias", [True, False])
 def test_unnest_w_no_table_references(faux_conn, alias):
     fcall = sqlalchemy.func.unnest([1, 2, 3])
@@ -444,14 +460,10 @@ def test_array_indexing(faux_conn, metadata):
         metadata,
         sqlalchemy.Column("a", sqlalchemy.ARRAY(sqlalchemy.String)),
     )
-    got = str(sqlalchemy.select([t.c.a[0]]).compile(faux_conn.engine))
+    got = str(sqlalchemy.select(t.c.a[0]).compile(faux_conn.engine))
     assert got == "SELECT `t`.`a`[OFFSET(%(a_1:INT64)s)] AS `anon_1` \nFROM `t`"
 
 
-@pytest.mark.skipif(
-    packaging.version.parse(sqlalchemy.__version__) < packaging.version.parse("1.4"),
-    reason="regexp_match support requires version 1.4 or higher",
-)
 def test_visit_regexp_match_op_binary(faux_conn):
     table = setup_table(
         faux_conn,
@@ -468,10 +480,6 @@ def test_visit_regexp_match_op_binary(faux_conn):
     assert result == expected
 
 
-@pytest.mark.skipif(
-    packaging.version.parse(sqlalchemy.__version__) < packaging.version.parse("1.4"),
-    reason="regexp_match support requires version 1.4 or higher",
-)
 def test_visit_not_regexp_match_op_binary(faux_conn):
     table = setup_table(
         faux_conn,
@@ -485,4 +493,67 @@ def test_visit_not_regexp_match_op_binary(faux_conn):
     result = sql_statement.compile(faux_conn).string
     expected = "NOT REGEXP_CONTAINS(`table`.`foo`, %(foo_1:STRING)s)"
 
+    assert result == expected
+
+
+def test_visit_mod_binary(faux_conn):
+    table = setup_table(
+        faux_conn,
+        "table",
+        sqlalchemy.Column("foo", sqlalchemy.Integer),
+    )
+    sql_statement = table.c.foo % 2
+    result = sql_statement.compile(faux_conn).string
+    expected = "MOD(`table`.`foo`, %(foo_1:INT64)s)"
+
+    assert result == expected
+
+
+def test_window_rows_between(faux_conn):
+    """This is a replacement for the
+    'test_window_rows_between'
+    test in sqlalchemy's suite of compliance tests.
+
+    Their test is expecting things in sorted order and BQ
+    doesn't return sorted results the way they expect so that
+    test fails.
+
+    Note: that test only appears in:
+    sqlalchemy/lib/sqlalchemy/testing/suite/test_select.py
+    in version 2.0.32. It appears as though that test will be
+    replaced with a similar but new test called:
+    'test_window_rows_between_w_caching'
+    due to the fact the rows are part of the cache key right now and
+    not handled as binds.  This is related to sqlalchemy Issue #11515
+
+    It is expected the new test will also have the same sorting failure.
+    """
+
+    table = setup_table(
+        faux_conn,
+        "table",
+        sqlalchemy.Column("id", sqlalchemy.String),
+        sqlalchemy.Column("col1", sqlalchemy.Integer),
+        sqlalchemy.Column("col2", sqlalchemy.Integer),
+    )
+
+    stmt = sqlalchemy.select(
+        sqlalchemy.func.max(table.c.col2).over(
+            order_by=[table.c.col1],
+            rows=(-5, 0),
+        )
+    )
+
+    sql = stmt.compile(
+        dialect=faux_conn.dialect,
+        compile_kwargs={"literal_binds": True},
+    )
+
+    result = str(sql)
+    expected = (
+        "SELECT max(`table`.`col2`) "
+        "OVER (ORDER BY `table`.`col1` "
+        "ROWS BETWEEN 5 PRECEDING AND CURRENT ROW) AS `anon_1` \n"  # newline character required here to match
+        "FROM `table`"
+    )
     assert result == expected